Can now orientate bridge skin lines parallel to the longest unsupported edge of the skin region.
The angle of the skin lines when bridging can make a big difference to the bridge quality.
In general, it is better for the skin lines to be parallel to skin edges that are unsupported.

This change detects the longest unsupported edge of a skin region and aligns the skin's
lines to be parallel to that edge.