Fix z-seam quality regression caused by recent commit.
Commit 61a234d5d fixed a problem that the optimizer would not align the z-seams correctly
in the specific situation where infill is printed before walls, the inner walls are
printed before the outer and the part has no holes.

That commit simply avoided using the optimizer when the part had no holes. Unfortunately,
that leads to more visible z-seams as the non-optimizing wall printing code doesn't try
to avoid retracting on the outer wall which the optimizer does. Therefore, this commit
reinstates using the optimizer when the part has no holes so the z-seam will be better
hidden again but it also avoids the previous problem by not executing the code that
fails when there are no holes.