Revert "Apply half line width offset if zig_zaggifying"
This reverts commit 2a8ff9f338d4d9e4c2959a2ed38d4ae2074cb4fc.

This only reverts parts of that commit which fit to the description.
Other type fixes and code style fixes remain intact.

The fix is reverted, because I fixed it already and this fix is wrong.
The outline_offset is defined as the offset from the polygons to get the outer boundary which to fill with the infill pattern.
It is therefore the Infill class which is responsible for applying an even further offset to get the outer wall of the infill.