Fix applying outline_offset+infill_overlap in two stages
When you offset first with outline_offset and then with infill_overlap, then you get a different result from when you offset once with outline_offset + infill_overlap, because the corners are tapered twice in the first case but only once in the second case. This could cause a crash because the crossings were indexed with a different number of line segments when it gathers the crossings than from when it connects them. So now we apply the offset once with outline_offset + infill_overlap to be the same as the original offset for finding crossings. This is essentially the sort of bugs that the TODO in the connecting algorithm tries to prevent.

Contributes to issue CURA-5361.