Fix no skin combing so that it doesn't cross skin areas!
The problem was that previously it considered the part's infill area(s) to define the
boundary of the combing area but then travel moves that didn't actually cross the infill
areas were considered to be OK to do with no retraction and so those travels could cross
skin areas!

So now, it adds the part's walls into the combing area so that any travel moves that cross
the part's outline will be combed correctly.