Now uses the shortest combing path around a polygon rather than the path with fewest points.
For most models this doesn't make a huge difference but for some travels it can when the
polygon outline point density is not uniform. As an example, think of a long thin rectangle with one end
square and the other end nicely rounded, the old code would always route via the square
end because it has fewer points. The new code will route using either end depending on which
direction yields the shortest path.

Slicing time doesn't appear to be unduly affected by this change.