Increase border around build volume to not go outside volume
If you had an extremely large branch diameter angle it would punch right through the border of 1mm and still print stuff outside. I've now increased the width of the border so it would have to be very very extreme to be able to print outside. You could still reproduce it if you set the angle to 89.999 degrees or something but this'll be fine for now.

Contributes to issue CURA-5037.