When lines are chained, don't leave singleton lines till after the chains have been printed.
Before, any lines that were not in a chain were printed after all the chains had been printed.
This would leave singleton lines (lines that are not joined to any others) dotted around
which could involve a fair amount of travel later to print. Now, when it looks for the next chain
to print, it also looks at the singleton lines and if any of those are closer than the
nearest chain end, it will print the singleton. The singleton lines are generally quite
short.