If possible, use the inside edge of the outer wall line as the combing boundary.
Before, it used the centre line of the outer wall as the outer boundary but it's better if
the outer boundary is inboard of the centre line to reduce the chance of dribble spoiling
the outer wall line so now we try and use the inner edge of the outer wall as the boundary.

It may be that the outer boundary can be moved even further inboard but that has not yet
been investigated.