Rewrite MergeInfillLines algorithm to work as a pre-pass
We want to allow this algorithm to actually remove lines from the list of paths. Because of that we must use it as a pre-filter to iterate in a separate loop over the paths rather than in-place altering the current line index that the main writing loop uses to write lines.

Contributes to issue CURA-5535.