Make getInnermostWalls() give a better result when the outer inset has gaps.
Before, when determining the outline of the 2nd inset, it noticed when there were gaps in
that inset and it used the 1st inset outline in those regions. Now, it does a similar job
when there are gaps in the 1st inset in which case it uses the part's outline in those
regions. This makes it possible to comb over thin walls which otherwise would not be
included in the combing region.