Instead of singleton, store Communication in Application
Singleton was not very feasible to implement, since the Communication class is virtual (so the static variable storing the class would need to become a pointer and you'd need to safely delete the instance at some point) and since the constructor of some implementations need parameters. Singletons with parameters are not feasible. So no singleton, just store it in Application and delete it upon the destructor of Application.

Contributes to issue CURA-4410.