When spiral is not being smoothed, maintain the alignment of the bulk of the first line segment.
Because the first line segment in a spiralized layer starts from the "wrong" location, its
direction will always be incorrect. If it is part of a flat surface, it is very obvious that the
line is not aligned with the layers above and below. So the workaround is to immediately
extrude from the start location (the seam vertex of the previous layer) to the closest point
on the line from the seam vertex of the current layer to the 2nd vertex. Once on the line,
the remainder of the line to the 2nd vertex will have the correct direction.

When this little jump onto the first line segment occurs close to a well defined model edge
it's pretty much invisible. It's more of a problem if the jump occurs in an area with a very
shallow slope.

More work required here but it may be a very difficult problem to really solve.