Partially revert "Simplify findSpiralizedLayerSeamVertexIndex()."
This reverts commit f623713df5737f2eedaa3089437419a4b8f5d800.

But some of that commit was good so this is a merge of the good bits and what we had before.

The problem with the new code was that it was too eager to move the seam back in the presence
of sharp corners whereas the previous code was OK because it took into account not just
the direction of the last segment of a layer but also the direction of the following segment.

Anyway, this combined commit appears to function OK.