Implement sliceNext()
This introduces compile errors though. That is because we're using methods to load settings from JSON documents which is still in the old format. We're also loading STL meshes using non-object-oriented programming which is just bad style. I'm about to change that. Also we're using printCall() from application and that should just remember the arguments from the last run() call so I don't need to provide it again here.

Contributes to issue CURA-4410.