Let setFirmwareDefaults accept Settings instance
Instead of SettingsBaseVirtual which is to be deprecated.
This entails that we're going to have to change the types of TimeEstimateCalculator's fields to the new, more specific types though. That has some consequences over the entire class and needs some overloads in the type definitions themselves.

Contributes to issue CURA-4410.