Obtain per-mesh settings from their settings classes in Mesh
We still have to do something about extruder index settings that are now obtained as size_t, while they might be -1. I'll look through the diff of this commit to get at those and obtain them via the ExtruderTrain& type instead, which does handle the -1 case.

Contributes to issue CURA-4410.