Get extruder nrs via ExtruderTrain
The size_t getter doesn't do anything fancy for any extruders if they have a negative extruder index. The ExtruderTrain& overload has a special handling for if the extruder train is -1, so that it'll get the extruder train of the current mesh. In any case, this is much safer against segfaults due to weird extruder indices.

Contributes to issue CURA-4410.