Divide ratio by 100 upon reading setting instead of constructor
We want the front-end to display this setting in percentages, so there it is multiplied by 100. But in CuraEngine we only want to use ratios. So when you construct a new Ratio instance we don't want to have to construct 100.0_r but use 1.0_r instead. So don't divide the ratio by 100 in the constructor, but divide it only when reading a setting.

Contributes to issue CURA-4410.