Add functions to get a setting as list of angles
And list of doubles.
The doubles is an adjustment of the previous call but with slightly broader regexes to allow radices in the numbers. The old int version can then call the doubles version as subroutine; it just needs to round. And the AngleDegrees version is just the doubles version but cast to AngleDegrees.

Contributes to issue CURA-4410.