Use Ratio for skin_density
This is one of the parameters that should really be determined inside processSkinPrintFeature, but because all the bridging code is in processTopBottom, we can't. All that bridging code should be in processSkinPrintFeature so that it also works with the roofing, but that is another refactor inside this refactor which I'm not about to start, and also changes the behaviour so it shouldn't be done now.

Contributes to issue CURA-4410.