Get travel_avoid_other_parts from extruder train in combing
Instead of passing this along in the parameters.
For rafts this results in a change in semantics, since previously this setting was passed along for the adhesion extruder train. Instead it's now taken from the first extruder for rafts as well (like it was for other layers previously). Luckily, the adhesion extruder should be the only extruder used on raft layers so it shouldn't make a difference in the actual g-code.

Contributes to issue CURA-4410.