Add decoy setting value to global stack
We don't want the limit to extruder to immediately fall back to the global stack. Previously this would result in a crash if it happened (since the global stack had no value for it and has no further inheritance). Now it results in a failed test.

Contributes to issue CURA-4410.