Fix support for higher parts extending below lower parts of the model.
part of CURA-5833

Support for higher parts would ignore parts of the model below and appear to
go through them (but not actually overlap) causing unnescesary and sometimes
hard to remove support. This fix intends to prevents that while also preserving
both the fix for CURA-5614 (fix for too much support removed in the X/Y
direction), while not confusing the support-stair calculations.