Re-enable 'prevent fat infill' fix for ironing as it solves CURA-5855 as well.
Ironing would be messed up becasue of spurious merging.
As ironing is part of PrintFeatureType::Skin, and we want the fix to be 'off'
for skin in general (as it prevents small 'arching' from merging into lines),
this meant that a way was needed to know if something is part of ironing
within the mergeInfillLines methods.

part of CURA-5776 and CURA-5885