Fix formula for checking maximum removed error
The error is meant to depict the distance squared between the vertex and the new line, not the area that gets removed. This implements that, but retains the property that the area/distance gets reduced by subsequent vertices that error in the other direction, so that the zigzag case should still work.

Contributes to issue CURA-5924.