CMake: Depend on Protobuf
When building the engine with Arcus, we are in the following lines generating C code using protoc.
However, if your installation is not complete and you've got all libs and headers installed, but no protoc installed, then the build fails.
At least that is the case on Windows, when protoc can't be found in %PATH%.

find_package should ensure here, that all libs, headers and command line tools are installed and found. If not this should throw an error and prevent the situation I ran into.