Rethink how the first and last vertex are handled in the list
Since the last vertex can't be checked for removal in the same loop, we handle it separately. This may mean that we need to re-add the first vertex (if it was removed) or that we need to check around the last vertex in order to see if it needs to be removed.
This does cause the test to have a different result. The 90 degree angle happens to bring the last vertex close to the first one again.

Contributes to issue CURA-5924.