Make prime-tower brim not overlap other skirt/brim.
Can't substract from the prime tower brim or the model brim, since the lines
are defined by solid polygons. This'll cause the concentric outlines on the
inside of a brim (or a skirt) to overlap each other in the area where avoidance
is needed.

Therefore, instead, pretend (_if_ both prime-tower-brim and model skirt/brim)
the prime-tower brim is part of the prime-tower for the purposes of the
generation of the other brim/skirt. Then, generate the prime-tower, including
its brim, in the spot left over.

This'll cause the brim-tower to have a sort of double-wide brim in the case
_both_ normal skirt-brim adhesion and the specific prime-tower brim are
enabled. You could argue that this is exactly what the user asked for though.

Part of [CURA-5864]