Don't make extruder_count a field of GCodeExport class
We can just request it again if necessary in every function. That saves memory and time when moving the class, and makes it more robust to changes in which stuff is constructed; e.g. constructing the export class before the setting is properly transmitted from the front-end. Not something we'll use now, but in general it's better to not duplicate the data when you don't need to.