Halve output of relativeHammingDistance
It makes less sense this way when looking at the algorithm semantically and the precise output. However when you look at this function from a usability point of view, it makes it easier if you return a value between 0 and 1 that just indicates how similar the two polygons are.

Contributes to issue CURA-6443.