Don't place support brim polygons on the back of skirt_brim as that upsets the ordering.
When the skirt/brim lines are printed, it is assumed that the polygon at the back of
skirt_brim is furthest away from the outline of the model. By adding support brim lines
to the back of skirt_brim, if any of those lines were inside of the model (i.e. support
inside a hole in the model) it would print those inner support brim lines first and then
go on to print the brim around the outside of the model from inner to outer (wrong!)

This change ensures that the support brim lines are before the original lines in skirt_brim
thus ensuring that the back polygon doesn't change.