Fix infinite loop with secondary extruder if platform adhesion is set to None
The previous fix was wrong because it causes the secondary extruders to not generate any skirt lines any more since their skirt_brim polygons started at 0. It happened to fix the loop because it skipped over the while loop, but the actual reason why the while loop was infinite is because first_layer_outline was empty, and that's the one that's being offset and so it's still empty.

Contributes to issues CURA-6641 and CURA-6647.