fix: discretize parabolas and line-line segments at boundary locations
this makes sure that the radial distance R is monotonic between two marked regions.
When going from one local maximum or marked region to the other,
one must pass a region where the curvature is low, so there should be a marked region in between.
However if we just use the discretization the significant regions might not get marked, so we introduce the right discretization which induces the marking