Fix support-brim missing if adhesion set to None.
A fix introduced an early-out for support-brim-generation if the polygons where of 0 size. However, support wasn't counted, and the (normal) polygons are set to empty if _only_ brim for support is generated and the adhesion if otherwise None. The check was moved downwards to where the infinite loop it fixed actually happens.
part of CURA-6700