Speed up the ordering of large numbers of polygons when using the combed travel distances.
Now, when there are more than 100 polygons, it simply increases the direct travel distance
by a factor of 5 when the direct travel hits the combing boundary. The intent is that travels
that hit the boundary are penalised in favour of travels that don't hit the boundary.

When there are 100 or less polygons, it will use the combed travel distances.