Don't use min XY distance when layer below protrudes far enough for support to be built on it.
This handles the situation where there is a sloping wall that has support clinging to it
at min XY distance meeting a flat model surface below with no z-gap. The bottom support
layer should use XY distance as long as the flat surface is big enough to have some support
on top.