Add tests that get executed in every possible combination of params
This is a much more maintainable and thorough way of testing this function. I can now add tests that test a certain property of the behaviour rather than having to go through all of the possible combinations that this code could run under.
The old tests can be removed now, but I'll add parameterised tests for the things that they are testing now and then remove them.

Contributes to issue CURA-6860.