fix test simplifyLimitedError
The previous test only considered single corners, while simplify() works cumulatively.
Even though one corner might not simplify based on its current geometry, it might be simplified if the previous corner was simplified.
This seems to be related to the fact that the spiral has self-intersectiong geometry.
The simplification of more segments can result in a smaller error because the spiral loops back unto itself.

In the new test the whole spiral is considered to be a region of small errors and the whole thing is simplifiable,
except that we added unsimplifiable geometry in order to prevent the collapse of the whole region,
which was another problem of the original test