fix simplify handling of colinear lines
The function loops over as many points as size(), so we don't need to handle the first point outside of the loop.

The code for checking colinearity was buggy and could escalate.
Now we simply interpret the function requirements as stating that the smallest_line_segment criterion
ca be ignored if the local deviation is smaller than 5 micron.
We therefore use the same vertex removing code for those small errors as for larger errors.