Revert "Merge pull request #1238 from Ultimaker/CURA-7352_fix_simple_blocker"
This reverts commit 683181c58be80c9a42862485e50dd8c0ea95b658, reversing
changes made to 3dd6a60725477e053b28e7c6840cd1e654cea504.

This reverts a cherry-pick that took a fix for oversimplifying the shape of supports from master into 4.6. The cherry-picked commit also had a bug that was later fixed on master, but that fix has further widespread impacts again leading to more risk of bugs later on.
Instead of that risk we decided to revert the fix made for CURA-7352. Support blockers might now be ineffective in some cases, but that seems to be a fairly rare case.

Contributes to issue CURA-7393.